- use correct data type   - int instead of byte - formatting - refactored - removed FileChooser   - deduce prc file from selected mp3

+++ src/main/java/cz/jcu/prf/uai/javamugs/clonehero/logic/Game.java
@@ -25,7 +25,7 @@ public class Game
    * @param difficulty game difficulty.
    * @param pressChart loaded press chart. Never null.
    */
-  public Game(double timeOffset, byte difficulty, PressChart pressChart)
+  public Game(double timeOffset, int difficulty, PressChart pressChart)
   {
     this.timeOffset = timeOffset;
     this.pressChart = pressChart;
